More stringent checks to determine if a frame object is valid. This check is used now in more palces. Also, some small refactoring was done to DRY code done in the process.
In an unpatched Ruby 1.9, the current threadframe behavior will mean
an infinite looping in raising an error rather than previously which
was a SEGV when the frame was accessed, I suspect there is a bug
possibly with respect to the way Ruby's handles errors raised by a
hook. I need to track this down soon.